fixed memory leaks. added interaction with webserver scripts and tested the results with tony.
turns out there's an issue with the order of items in the xml list, however, the app code for upload and download and script activation has proved to work.

if it doesn't work for you, presently it is because of a problem with the php script. 

it may work better to use the device GUID instead of email to identify each profile.